% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfidd-stan-tools.r
\name{.nfidd_source_stan_path}
\alias{.nfidd_source_stan_path}
\title{Find the Stan code in a package source tree}
\usage{
.nfidd_source_stan_path(path = getwd())
}
\arguments{
\item{path}{Character string, the directory to start from.}
}
\value{
A character string with the path to `inst/stan`, or `NULL`.
}
\description{
Walks up from `path` looking for the source tree of this package, i.e. a
`DESCRIPTION` naming `nfidd.nowcasting` next to an `inst/stan` directory.
Returns `NULL` when there is no such directory above `path`.
}
\keyword{internal}
