% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfidd-stan-tools.r
\name{nfidd_stan_models}
\alias{nfidd_stan_models}
\title{List Available Stan Models in NFIDD}
\usage{
nfidd_stan_models(stan_path = nfidd.nowcasting::nfidd_stan_path())
}
\arguments{
\item{stan_path}{Character vector of directories holding Stan files, as
returned by [nfidd_stan_path()]. May name more than one directory.}
}
\value{
A character vector of available Stan model names.
}
\description{
This function finds all available Stan models in the NFIDD package and
returns their names without the .stan extension.
}
\examples{
nfidd_stan_models()
}
